SDK-537: Move YotiClient tests to separate file
